pyproject.toml
[project]
name = "multimodel-mcp"
version = "0.1.0"
description = "Multimodel MCP server with file system operations and Gemini AI integration"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp-inspector>=0.1.0",
    "mcp-use>=1.4.1",
    "mcp[cli]>=1.22.0",
    "nest-asyncio>=1.6.0",
    "langchain-google-genai>=1.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
]
